Chore: Requirement, WorkList 열거형 파일 분리 후 폴더링
